# Optimize URL bar history and bookmark suggestions (generateNewSuggestionsList())