Related to #7453

With 10,000 sites, time to run `generateNewSuggestionsList()`:

- Autocomplete history and bookmarks **Enabled**:
  - Existing (33e3fca1025ff91e656567c0ad4e862ecc312000): 170 ms
  - New: 38 ms
- Autocomplete history and bookmarks **Disabled**:
  - Existing: 120 ms
  - New: 3 ms

Auditors: @darkdh @bbondy

- [x] Submitted a [ticket](https://github.com/brave/browser-laptop/issues) for my issue if one did not already exist.
- [ ] Used Github [auto-closing keywords](https://help.github.com/articles/closing-issues-via-commit-messages/) in the commit message.
- [x] Added/updated tests for this change (for new code or code which already has tests).
- [x] Ran `git rebase -i` to squash commits (if needed).

Test Plan:
- Run tests: `npm run test -- --grep="^urlBarSuggestions"`